requirements.txt
# Core dependencies
streamlit>=1.28.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
openai>=1.10.0

# API clients
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.8.0

# Optional: For better error handling and logging
tenacity>=8.2.0

# File watching (e.g. for Streamlit reload)
watchdog>=6.0.0